+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - User Messages Tool now lists tick and time.
 - Changes
     - User Messages Tool ui now behaves closer to frames/game events too, closes #12
+    - consolidated various table widths, closes #14
 - Internal
     - set source-demo-tool version to 0.5.0